Refactor to use wallet type
